more directory restructuring and deleting old files we better just redo
