+++ src/app/component/page/about-page/about-page.component.html
@@ -1,0 +1,1 @@
+<app-about-section></app-about-section>
