+++ src/app/component/page/faq-page/faq-page.component.ts
