+++ src/app/component/page/news-page/news-page.component.ts
