+++ src/app/component/section/about-section/about-section.component.html
@@ -1,7 +1,6 @@
 <section class="section" id="about-section">
   <div class="section-content">
     <div class="grid">
-
       <div class="column">
         <div class="column">
           <div class="column two">
