+++ src/app/component/section/about-section/about-section.component.scss
