+++ src/app/component/section/landing-section/landing-section.component.scss
