+++ src/app/component/section/news-section/news-section.component.ts
