--- src/app/module/plea/component/haf-card/haf-card.component.html
@@ -1,53 +1,0 @@
-<!--<div class="card">-->
-<!--  <div class="card-block c-header">-->
-<!--    <h4 class="card-title">{{ plea.veganProduct!.name }}</h4>-->
-<!--    <span class="card-subtitle">-->
-<!--      <clr-icon shape="crown" size="22"></clr-icon>-->
-<!--      by {{ plea.company.name }}</span-->
-<!--    >-->
-<!--  </div>-->
-
-<!--  &lt;!&ndash;  will be vegan product productImage&ndash;&gt;-->
-<!--  <div class="card-img">-->
-<!--    <img src="/assets/images/{{ plea.nonVeganProduct.imageUrl }}" />-->
-<!--  </div>-->
-
-<!--  <div class="card-block">-->
-<!--    <div class="card-body">-->
-<!--      <p>-->
-<!--        <span class="highlighted-txt">{{ plea.company.name }}</span> has read our pleas and on-->
-<!--        <span class="highlighted-txt">date</span> they replaced-->
-<!--        <span class="highlighted-txt">{{ plea.nonVeganProduct.name }}</span> with this new vegan product!-->
-<!--      </p>-->
-
-<!--      <p>-->
-<!--        <span class="highlighted-txt">{{ plea.company.name }}</span-->
-<!--        >, on behalf of all animals and us Pleagans, thank you very much!-->
-<!--      </p>-->
-
-<!--      <p>-->
-<!--        Be sure to support their product and leave a little comment below to show them your appreciation!-->
-<!--        <a class="underlined" [routerLinkActive]="'active'" [routerLink]="['/', 'hall-of-fame', 'details', plea.id]">-->
-<!--          Find further product details here</a-->
-<!--        >-->
-<!--        and spread the word in the your social circles!-->
-<!--      </p>-->
-<!--    </div>-->
-<!--  </div>-->
-
-<!--  <div class="card-block c-footer">-->
-<!--    <a class="btn btn-primary btn-icon">-->
-<!--      <clr-icon shape="heart" size="16"></clr-icon>-->
-<!--      Send some love-->
-<!--    </a>-->
-<!--    &lt;!&ndash;news post link goes here, with info on where to buy&ndash;&gt;-->
-<!--    <a-->
-<!--      class="btn btn-primary btn-icon"-->
-<!--      [routerLinkActive]="'active'"-->
-<!--      [routerLink]="['/', 'hall-of-fame', 'details', plea.id]"-->
-<!--    >-->
-<!--      <clr-icon shape="details" size="16"></clr-icon>-->
-<!--      Product details-->
-<!--    </a>-->
-<!--  </div>-->
-<!--</div>-->
