--- src/app/module/plea/component/haf-details/haf-details.component.html
@@ -1,1 +1,0 @@
-<!--<p>product works!</p>-->
