--- src/app/module/plea/component/haf-details/haf-details.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-product',
-  templateUrl: './haf-details.component.html',
-  styleUrls: ['./haf-details.component.scss'],
-})
-export class HafDetailsComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
