--- src/app/module/plea/component/hall-of-fame/hall-of-fame.component.html
@@ -1,3 +1,0 @@
-<!--TODO: rename component: Spotlight-->
-<!--path is just: /plea/complied-->
-<router-outlet></router-outlet>
