--- src/app/module/plea/component/hall-of-fame/hall-of-fame.component.ts
@@ -1,12 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-hall-of-fame',
-  templateUrl: './hall-of-fame.component.html',
-  styleUrls: ['./hall-of-fame.component.scss'],
-})
-export class HallOfFameComponent implements OnInit {
-  constructor() {}
-
-  ngOnInit(): void {}
-}
